refactor: move account to shard uid to EpochManagerHandle
