+++ chain/chain/src/types.rs
@@ -343,17 +343,8 @@ pub trait RuntimeAdapter: EpochManagerAdapter + Send + Sync {
 
     fn num_data_parts(&self) -> usize;
 
-    /// Account Id to Shard Id mapping, given current number of shards.
-    fn account_id_to_shard_id(
-        &self,
-        account_id: &AccountId,
-        epoch_id: &EpochId,
-    ) -> Result<ShardId, Error>;
-
     /// Returns `account_id` that suppose to have the `part_id`.
     fn get_part_owner(&self, epoch_id: &EpochId, part_id: u64) -> Result<AccountId, Error>;
-
-    fn shard_id_to_uid(&self, shard_id: ShardId, epoch_id: &EpochId) -> Result<ShardUId, Error>;
 
     /// Returns true if the shard layout will change in the next epoch
     /// Current epoch is the epoch of the block after `parent_hash`
